Web Text Downloader 2.0
Web Text Downloader is a program for automatically extracting and
storing predetermined content from both open access and
password-protected websites. The motivation to create this software
emerged from the need to extract a large number of news reports from an
on-line news report repository that recurrently requested entering the
username and password information. This software is an attempt to reduce
the burden of such tedious effort. Web Text Downloader extracts web
content that then can be processed with Web2Eventus to create the corpus
of text using for event coding in Eventus ID.